Update UserController to validate user from token

+++ src/main/java/co/booleanuk/bargain/controllers/UserController.java
@@ -4,9 +4,12 @@ import co.booleanuk.bargain.models.User;
 import co.booleanuk.bargain.models.responses.MessageResponse;
 import co.booleanuk.bargain.services.UserService;
 import lombok.AllArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.web.bind.annotation.*;
 
@@ -14,6 +17,7 @@ import java.time.ZonedDateTime;
 import java.util.Optional;
 import java.util.UUID;
 
+@Slf4j
 @RestController
 @RequestMapping("user")
 @AllArgsConstructor
@@ -29,10 +33,16 @@ public class UserController {
 
     @PutMapping("{id}")
     public ResponseEntity<?> update(@PathVariable UUID id, @RequestBody User user) {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        String currentUserEmail = authentication.getName();
+
         Optional<User> userToUpdateOptional = this.userService.getUserById(id);
 
         if (userToUpdateOptional.isEmpty()) {
             return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+        }
+        if (!currentUserEmail.equals(userToUpdateOptional.get().getEmail())) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
         }
 
         User userToUpdate = userToUpdateOptional.get();
@@ -51,20 +61,23 @@ public class UserController {
 
     @DeleteMapping("{id}")
     public ResponseEntity<?> deleteById(@PathVariable UUID id) {
-        Optional<User> userToDeleteOptional = this.userService.getUserById(id);
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        String currentUserEmail = authentication.getName();
 
+        Optional<User> userToDeleteOptional = this.userService.getUserById(id);
         if (userToDeleteOptional.isEmpty()) {
             return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
         }
-
         User userToDelete = userToDeleteOptional.get();
+        if (!currentUserEmail.equals(userToDelete.getEmail())) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
 
         try {
             this.userService.deleteUser(userToDelete);
         } catch (DataIntegrityViolationException e) {
             return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(new MessageResponse("Error: Violation of foreign key constraint"));
         }
-
         return ResponseEntity.ok(userToDelete);
     }
 }
